+++ src/controllers/AssetsController.php
@@ -21,11 +21,9 @@ use craft\fields\Assets as AssetsField;
 use craft\helpers\App;
 use craft\helpers\Assets;
 use craft\helpers\Db;
-use craft\helpers\Image;
 use craft\helpers\ImageTransforms;
 use craft\helpers\StringHelper;
 use craft\i18n\Formatter;
-use craft\image\Raster;
 use craft\imagetransforms\ImageTransformer;
 use craft\models\VolumeFolder;
 use craft\web\Controller;
@@ -760,13 +758,14 @@ class AssetsController extends Controller
         $assetId = (int)$this->request->getRequiredQueryParam('assetId');
         $size = (int)$this->request->getRequiredQueryParam('size');
 
-        $filePath = Assets::getImageEditorSource($assetId, $size);
-
-        if (!$filePath) {
+        $asset = Asset::findOne($assetId);
+        if (!$asset) {
             throw new BadRequestHttpException('The Asset cannot be found');
         }
 
-        return $this->response->sendFile($filePath, null, ['inline' => true]);
+        $url = Craft::$app->getImageTransforms()->getSimpleTransformUrlForAsset($asset, $size, $size, 'fit');
+
+        return $this->response->redirect($url);
     }
 
     /**
@@ -822,36 +821,21 @@ class AssetsController extends Controller
                 throw new BadRequestHttpException('Invalid cropping parameters passed');
             }
 
+            // TODO Fire an event for any other image editing takers.
+            $transformer = new ImageTransformer();
+
+            $originalImageWidth = $asset->width;
+            $originalImageHeight = $asset->height;
+
+            $transformer->startImageEditing($asset);
+
             $imageCropped = ($cropData['width'] !== $imageDimensions['width'] || $cropData['height'] !== $imageDimensions['height']);
             $imageRotated = $viewportRotation !== 0 || $imageRotation !== 0.0;
             $imageFlipped = !empty($flipData['x']) || !empty($flipData['y']);
             $imageChanged = $imageCropped || $imageRotated || $imageFlipped;
 
-            $imageCopy = $asset->getCopyOfFile();
-
-            $imageSize = Image::imageSize($imageCopy);
-
-            /** @var Raster $image */
-            $image = Craft::$app->getImages()->loadImage($imageCopy, true, max($imageSize));
-
-            // TODO Is this hacky? It seems hacky.
-            // We're rasterizing SVG, we have to make sure that the filename change does not get lost
-            if (strtolower($asset->getExtension()) === 'svg') {
-                unlink($imageCopy);
-                $imageCopy = preg_replace('/(svg)$/i', 'png', $imageCopy);
-                $asset->setFilename(preg_replace('/(svg)$/i', 'png', $asset->getFilename()));
-            }
-
-            [$originalImageWidth, $originalImageHeight] = $imageSize;
-
             if ($imageFlipped) {
-                if (!empty($flipData['x'])) {
-                    $image->flipHorizontally();
-                }
-
-                if (!empty($flipData['y'])) {
-                    $image->flipVertically();
-                }
+                $transformer->flipImage(!empty($flipData['x']), !empty($flipData['y']));
             }
 
             $generalConfig = Craft::$app->getConfig()->getGeneral();
@@ -859,17 +843,17 @@ class AssetsController extends Controller
             $generalConfig->upscaleImages = true;
 
             if ($zoom !== 1.0) {
-                $image->scaleToFit($originalImageWidth * $zoom, $originalImageHeight * $zoom);
+                $transformer->scaleImage($originalImageWidth * $zoom, $originalImageHeight * $zoom);
             }
 
             $generalConfig->upscaleImages = $upscale;
 
             if ($imageRotated) {
-                $image->rotate($imageRotation + $viewportRotation);
+                $transformer->rotateImage($imageRotation + $viewportRotation);
             }
 
-            $imageCenterX = $image->getWidth() / 2;
-            $imageCenterY = $image->getHeight() / 2;
+            $imageCenterX = $transformer->getEditedImageWidth() / 2;
+            $imageCenterY = $transformer->getEditedImageHeight() / 2;
 
             $adjustmentRatio = min($originalImageWidth / $imageDimensions['width'], $originalImageHeight / $imageDimensions['height']);
             $width = $cropData['width'] * $zoom * $adjustmentRatio;
@@ -891,11 +875,13 @@ class AssetsController extends Controller
             }
 
             if ($imageCropped) {
-                $image->crop($x, $x + $width, $y, $y + $height);
+                $transformer->crop($x, $y, $width, $height);
             }
 
             if ($imageChanged) {
-                $image->saveAs($imageCopy);
+                $finalImage = $transformer->finishImageEditing();
+            } else {
+                $finalImage = $transformer->cancelImageEditing();
             }
 
             $output = [];
@@ -912,7 +898,7 @@ class AssetsController extends Controller
 
                 // Only replace file if it changed, otherwise just save changed focal points
                 if ($imageChanged) {
-                    $assets->replaceAssetFile($asset, $imageCopy, $asset->getFilename());
+                    $assets->replaceAssetFile($asset, $finalImage, $asset->getFilename());
                 } else if ($focalChanged) {
                     Craft::$app->getElements()->saveElement($asset);
                 }
@@ -921,7 +907,7 @@ class AssetsController extends Controller
                 $newAsset->avoidFilenameConflicts = true;
                 $newAsset->setScenario(Asset::SCENARIO_CREATE);
 
-                $newAsset->tempFilePath = $imageCopy;
+                $newAsset->tempFilePath = $finalImage;
                 $newAsset->setFilename($asset->getFilename());
                 $newAsset->newFolderId = $folder->id;
                 $newAsset->setVolumeId($folder->volumeId);
@@ -1020,17 +1006,12 @@ class AssetsController extends Controller
         }
 
         try {
-            $path = Craft::$app->getAssets()->getThumbPath($asset, $width, $height, true);
+            $url = Craft::$app->getImageTransforms()->getSimpleTransformUrlForAsset($asset, $width, $height);
+            return $this->response->redirect($url);
         } catch (Throwable $e) {
             Craft::$app->getErrorHandler()->logException($e);
             return $this->asBrokenImage($e);
         }
-
-        return $this->response
-            ->setCacheHeaders()
-            ->sendFile($path, $asset->getFilename(), [
-                'inline' => true,
-            ]);
     }
 
     /**