+++ src/imagetransforms/ImageTransformer.php
@@ -10,6 +10,7 @@ namespace craft\imagetransforms;
 
 use Craft;
 use craft\base\imagetransforms\EagerImageTransformerInterface;
+use craft\base\imagetransforms\ImageEditorTransformerInterface;
 use craft\base\imagetransforms\ImageTransformerInterface;
 use craft\base\LocalFsInterface;
 use craft\db\Query;
@@ -43,12 +44,17 @@ use yii\base\InvalidConfigException;
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 4.0.0
  */
-class ImageTransformer implements ImageTransformerInterface, EagerImageTransformerInterface
+class ImageTransformer implements ImageTransformerInterface, EagerImageTransformerInterface, ImageEditorTransformerInterface
 {
     /**
      * @var ImageTransformIndex[]
      */
     protected array $eagerLoadedTransformIndexes = [];
+
+    /**
+     * @var array
+     */
+    protected array $imageEditorData = [];
 
     /**
      * Returns the URL for an image asset transform.
@@ -64,7 +70,7 @@ class ImageTransformer implements ImageTransformerInterface, EagerImageTransform
         $imageTransformIndex = $this->getTransformIndex($asset, $imageTransform);
 
         if ($imageTransformIndex->fileExists) {
-            $fs = $asset->getFs();
+            $fs = $asset->getVolume()->getTransformFs();
             $uri = $this->getTransformUri($asset, $imageTransformIndex);
 
             // Check if it really exists
@@ -72,7 +78,7 @@ class ImageTransformer implements ImageTransformerInterface, EagerImageTransform
                 $imageTransformIndex->fileExists = false;
                 $this->storeTransformIndexData($imageTransformIndex);
             } else {
-                return AssetsHelper::generateUrl($asset->getVolume(), $asset, $uri, $imageTransformIndex->dateUpdated);
+                return AssetsHelper::generateUrl($fs, $asset, $uri, $imageTransformIndex->dateUpdated);
             }
         }
 
@@ -100,7 +106,7 @@ class ImageTransformer implements ImageTransformerInterface, EagerImageTransform
      */
     protected function deleteImageTransform(Asset $asset, ImageTransformIndex $transformIndex): void
     {
-        $asset->getFs()->deleteFile($asset->folderPath . $this->getTransformSubpath($asset, $transformIndex));
+        $asset->getVolume()->getTransformFs()->deleteFile($asset->folderPath . $this->getTransformSubpath($asset, $transformIndex));
     }
 
     /**
@@ -289,10 +295,11 @@ class ImageTransformer implements ImageTransformerInterface, EagerImageTransform
 
         $volume = $asset->getVolume();
         $transformPath = $asset->folderPath . $this->getTransformSubpath($asset, $index);
+        $transformFs = $volume->getTransformFs();
 
         // Already created. Relax, grasshopper!
-        if ($volume->getFs()->fileExists($transformPath)) {
-            $dateModified = $volume->getFs()->getDateModified($transformPath);
+        if ($transformFs->fileExists($transformPath)) {
+            $dateModified = $transformFs->getDateModified($transformPath);
             $parameterChangeTime = $index->getTransform()->parameterChangeTime;
 
             if (!$parameterChangeTime || $parameterChangeTime->getTimestamp() <= $dateModified) {
@@ -345,28 +352,16 @@ class ImageTransformer implements ImageTransformerInterface, EagerImageTransform
             $image->setInterlace($transform->interlace);
         }
 
-        $event = new GenerateTransformEvent([
-            'transformIndex' => $index,
-            'asset' => $asset,
-            'image' => $image,
-        ]);
-
-        Craft::$app->getImageTransforms()->trigger(ImageTransforms::EVENT_GENERATE_TRANSFORM, $event);
-
-        if ($event->tempPath !== null) {
-            $tempPath = $event->tempPath;
-        } else {
-            $tempFilename = uniqid(pathinfo($index->filename, PATHINFO_FILENAME), true) . '.' . $index->detectedFormat;
-            $tempPath = Craft::$app->getPath()->getTempPath() . DIRECTORY_SEPARATOR . $tempFilename;
-            $image->saveAs($tempPath);
-        }
+        $tempFilename = uniqid(pathinfo($index->filename, PATHINFO_FILENAME), true) . '.' . $index->detectedFormat;
+        $tempPath = Craft::$app->getPath()->getTempPath() . DIRECTORY_SEPARATOR . $tempFilename;
+        $image->saveAs($tempPath);
 
         clearstatcache(true, $tempPath);
 
         $stream = fopen($tempPath, 'rb');
 
         try {
-            $volume->getFs()->writeFileFromStream($transformPath, $stream, []);
+            $volume->getTransformFs()->writeFileFromStream($transformPath, $stream, []);
         } catch (FsException $e) {
             Craft::$app->getErrorHandler()->logException($e);
         }
@@ -396,7 +391,7 @@ class ImageTransformer implements ImageTransformerInterface, EagerImageTransform
         $index->filename = $transformFilename;
 
         $matchFound = $this->getSimilarTransformIndex($asset, $index);
-        $fs = $volume->getFs();
+        $fs = $volume->getTransformFs();
 
         // If we have a match, copy the file.
         if ($matchFound) {
@@ -674,6 +669,104 @@ class ImageTransformer implements ImageTransformerInterface, EagerImageTransform
 
         return $result ? new ImageTransformIndex($result) : null;
     }
+
+    /**
+     * @inheritdoc
+     */
+    public function startImageEditing(Asset $asset): void
+    {
+        $imageCopy = $asset->getCopyOfFile();
+
+        /** @var Raster $image */
+        $image = Craft::$app->getImages()->loadImage($imageCopy, true, max($asset->height, $asset->width));
+
+        // TODO Is this hacky? It seems hacky.
+        // We're rasterizing SVG, we have to make sure that the filename change does not get lost
+        if (strtolower($asset->getExtension()) === 'svg') {
+            unlink($imageCopy);
+            $imageCopy = preg_replace('/(svg)$/i', 'png', $imageCopy);
+            $asset->setFilename(preg_replace('/(svg)$/i', 'png', $asset->getFilename()));
+        }
+
+        $this->imageEditorData['image'] = $image;
+        $this->imageEditorData['tempLocation'] = $imageCopy;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function flipImage(bool $flipX, bool $flipY): void
+    {
+        if ($flipX) {
+            $this->imageEditorData['image']->flipHorizontally();
+        }
+        if ($flipY) {
+            $this->imageEditorData['image']->flipVertically();
+        }
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function scaleImage(int $width, int $height): void
+    {
+        $this->imageEditorData['image']->scaleToFit($width, $height);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function rotateImage(float $degrees): void
+    {
+        $this->imageEditorData['image']->rotate($degrees);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getEditedImageWidth(): int
+    {
+        return $this->imageEditorData['image']->getWidth();
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getEditedImageHeight(): int
+    {
+        return $this->imageEditorData['image']->getHeight();
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function crop(int $x, int $y, int $width, int $height): void
+    {
+        $this->imageEditorData['image']->crop($x, $x + $width, $y, $y + $height);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function finishImageEditing(): string
+    {
+        $tempLocation = $this->imageEditorData['tempLocation'];
+        $this->imageEditorData['image']->saveAs($tempLocation);
+        $this->imageEditorData = [];
+
+        return $tempLocation;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function cancelImageEditing(): string
+    {
+        $tempLocation = $this->imageEditorData['tempLocation'];
+        $this->imageEditorData = [];
+        return $tempLocation;
+    }
+
 
     /**
      * Delete transform records by an Asset id