+++ src/services/ImageTransforms.php
@@ -18,9 +18,10 @@ use craft\db\Table;
 use craft\elements\Asset;
 use craft\errors\ImageTransformException;
 use craft\events\AssetEvent;
+use craft\events\AssetGetThumbnailUrlEvent;
 use craft\events\ConfigEvent;
 use craft\events\ImageTransformEvent;
-use craft\events\RegisterImageTransformersEvent;
+use craft\events\RegisterComponentTypesEvent;
 use craft\helpers\Assets as AssetsHelper;
 use craft\helpers\Db;
 use craft\helpers\FileHelper;
@@ -75,9 +76,9 @@ class ImageTransforms extends Component
     public const EVENT_AFTER_DELETE_IMAGE_TRANSFORM = 'afterDeleteImageTransform';
 
     /**
-     * @event GenerateTransformEvent The event that is triggered when a transform is being generated for an Asset.
+     * @event AssetGetThumbnailUrlEvent The event that is triggered when a transform is being generated for an Asset.
      */
-    public const EVENT_GENERATE_TRANSFORM = 'generateTransform';
+    public const EVENT_GET_THUMB_URL = 'getThumbUrl';
 
     /**
      * @event AssetEvent The event that is triggered when a transform is being generated for an Asset.
@@ -85,7 +86,7 @@ class ImageTransforms extends Component
     public const EVENT_BEFORE_INVALIDATE_ASSET_TRANSFORMS = 'beforeInvalidateAssetTransforms';
 
     /**
-     * @event RegisterImageTransformersEvent The event that is triggered when registering image transformers.
+     * @event RegisterComponentTypesEvent The event that is triggered when registering image transformers.
      */
     public const EVENT_REGISTER_IMAGE_TRANSFORMERS = 'registerImageTransformers';
 
@@ -470,9 +471,30 @@ class ImageTransforms extends Component
     }
 
     /**
-     * @param string $type
+     * Get all image transformer types.
+     *
+     * @return array
+     */
+    public function getAllImageTransformerTypes(): array
+    {
+        $transformerTypes = [
+            ImageTransformer::class,
+        ];
+
+        $event = new RegisterComponentTypesEvent([
+            'types' => $transformerTypes,
+        ]);
+
+        $this->trigger(self::EVENT_REGISTER_IMAGE_TRANSFORMERS, $event);
+
+        return $event->types;
+    }
+
+    /**
+     * @template T
+     * @param class-string<T> $type
      * @param array $config
-     * @return ImageTransformerInterface
+     * @return T
      * @throws InvalidConfigException
      */
     public function getImageTransformer(string $type, array $config = []): ImageTransformerInterface
@@ -486,6 +508,34 @@ class ImageTransforms extends Component
         }
 
         return $this->_imageTransformers[$type];
+    }
+
+    /**
+     * Return the URL for a simple transform. A simple transform only allows height/width modifications.
+     *
+     * @param Asset $asset
+     * @param int $width
+     * @param int $height
+     * @param string $mode
+     * @return string
+     */
+    public function getSimpleTransformUrlForAsset(Asset $asset, int $width, int $height, string $mode = 'crop'): string
+    {
+        $event = new AssetGetThumbnailUrlEvent([
+            'asset' => $asset,
+            'transformer' => ImageTransformer::class
+        ]);
+
+        $this->trigger(self::EVENT_GET_THUMB_URL, $event);
+
+        $transform = new ImageTransform([
+            'width' => $width,
+            'height' => $height,
+            'mode' => $mode,
+            'transformer' => $event->transformer
+        ]);
+
+        return $transform->getImageTransformer()->getTransformUrl($asset, $transform, true);
     }
 
     /**
@@ -565,16 +615,16 @@ class ImageTransforms extends Component
     public function getAllImageTransformers(): array
     {
         $transformers = [
-            ImageTransformer::class
+            ImageTransformer::class,
         ];
 
-        $event = new RegisterImageTransformersEvent([
-            'transformers' => $transformers,
+        $event = new RegisterComponentTypesEvent([
+            'types' => $transformers,
         ]);
 
         $this->trigger(self::EVENT_REGISTER_IMAGE_TRANSFORMERS, $event);
 
-        return $event->transformers;
+        return $event->types;
     }
 
     /**