+++ src/services/Volumes.php
@@ -349,6 +349,7 @@ class Volumes extends Component
             $volumeRecord->name = $data['name'];
             $volumeRecord->handle = $data['handle'];
             $volumeRecord->fs = $data['fs'];
+            $volumeRecord->transformFs = $data['transformFs'];
             $volumeRecord->sortOrder = $data['sortOrder'];
             $volumeRecord->titleTranslationMethod = $data['titleTranslationMethod'] ?? Field::TRANSLATION_METHOD_SITE;
             $volumeRecord->titleTranslationKeyFormat = $data['titleTranslationKeyFormat'] ?? null;
@@ -638,6 +639,7 @@ class Volumes extends Component
                 'name',
                 'handle',
                 'fs',
+                'transformFs',
                 'titleTranslationMethod',
                 'titleTranslationKeyFormat',
                 'sortOrder',