+++ CHANGELOG.md
@@ -91,6 +91,7 @@
 - Added `craft\base\Image::heartbeat()`.
 - Added `craft\base\Image::setHeartbeatCallback()`.
 - Added `craft\base\imagetransforms\EagerImageTransformerInterface`.
+- Added `craft\base\imagetransforms\ImageEditorTransformerInterface`.
 - Added `craft\base\imagetransforms\ImageTransformerInterface`.
 - Added `craft\base\LocalFsInterface`.
 - Added `craft\base\NameTrait`.
@@ -177,7 +178,6 @@
 - Added `craft\events\DefineHtmlEvent::$static`.
 - Added `craft\events\ImageTransformEvent`.
 - Added `craft\events\RegisterConditionRuleTypesEvent`.
-- Added `craft\events\RegisterImageTransformersEvent`.
 - Added `craft\events\TransformImageEvent`.
 - Added `craft\fieldlayoutelements\addresses\AddressField`.
 - Added `craft\fieldlayoutelements\addresses\CountryCodeField`.
@@ -426,6 +426,8 @@
 - `craft\elements\Asset::getVolume()` now returns an instance of `craft\models\Volume`.
 - `craft\elements\db\ElementQuery::ids()` no longer accepts an array of criteria params.
 - `craft\events\DraftEvent::$source` has been renamed to `$canonical`.
+- `craft\events\DefineAssetThumbUrlEvent` now has a `$transform` field, instead of separate `$width` and `$height` fields.
+- `craft\events\GenerateTransformEvent` now has the `$asset`, `$transform` and `$url` fields.
 - `craft\events\GetAssetThumbUrlEvent` has been renamed to `DefineAssetThumbUrlEvent`.
 - `craft\events\GetAssetUrlEvent` has been renamed to `DefineAssetUrlEvent`.
 - `craft\events\RevisionEvent::$source` has been renamed to `$canonical`.
@@ -446,7 +448,9 @@
 - `craft\helpers\App::env()` now returns `null` if a value couldn’t be found, rather than `false`.
 - `craft\helpers\App::env()` now returns a boolean if the original value was `'true'` or `'false'`.
 - `craft\helpers\App::env()` now returns an integer or float if the original value was numeric.
+- `craft\helpers\Assets::generateUrl()` no longer accepts a `craft\models\Volume` object as the first argument. An object implementing `craft\base\FsInterface` is expected instead.
 - `craft\helpers\Assets::generateUrl()` no longer accepts a transform index for date modified comparisons. A `DateTime` object is expected instead.
+- `craft\helpers\Assets::generateUrl()` no longer accepts a `craft\models\Volume` as the first argument. A `craft\elements\Asset` object is expected instead.
 - `craft\helpers\Assets::urlAppendix()` no longer accepts a transform index for date modified comparisons. A `DateTime` object is expected instead.
 - `craft\helpers\Db::batchInsert()`, `craft\helpers\Db::insert()`, `craft\db\Command::batchInsert()`, `craft\db\Command::insert()`, `craft\db\Migration::batchInsert()`, and `craft\db\Migration::insert()` no longer have `$includeAuditColumns` arguments, and now check if the table has `dateCreated`, `dateUpdated`, and/or `uid` columns before setting their values.
 - `craft\helpers\Db::parseParam()` now validates that numeric values are passed if the `$columnType` is set to a numeric column type. ([#9142](https://github.com/craftcms/cms/issues/9142))